finish adding check_estimator to all transformers
