+++ feature_engine/selection/recursive_feature_addition.py
@@ -1,7 +1,6 @@
 from typing import List, Union
 
 import pandas as pd
-from sklearn.base import clone
 from sklearn.model_selection import cross_validate
 
 from feature_engine.dataframe_checks import _is_dataframe
@@ -84,9 +83,6 @@ class RecursiveFeatureAddition(BaseSelector):
     variables_:
         The variables that were be evaluated
 
-    estimator_;
-        a clone of the estimator
-
     n_features_in:
         The number of features in the train set used in fit
 
@@ -146,11 +142,9 @@ class RecursiveFeatureAddition(BaseSelector):
         # find numerical variables or check variables entered by user
         self.variables_ = _find_or_check_numerical_variables(X, self.variables)
 
-        self.estimator_ = clone(self.estimator)
-
         # train model with all features and cross-validation
         model = cross_validate(
-            self.estimator_,
+            self.estimator,
             X[self.variables_],
             y,
             cv=self.cv,
@@ -187,7 +181,7 @@ class RecursiveFeatureAddition(BaseSelector):
 
         # Run baseline model using only the most important feature
         baseline_model = cross_validate(
-            self.estimator_,
+            self.estimator,
             X[first_most_important_feature].to_frame(),
             y,
             cv=self.cv,
