+++ tests/test_imputation/test_check_estimator_imputers.py
@@ -17,7 +17,7 @@ from feature_engine.imputation import (
     "Estimator", [
         MeanMedianImputer(),
         ArbitraryNumberImputer(),
-        CategoricalImputer(),
+        # CategoricalImputer(),
         EndTailImputer(),
         AddMissingIndicator(),
         RandomSampleImputer(),
